songcrud/requirements.txt
asgiref==3.5.2
confusable-homoglyphs==3.2.0
distlib==0.3.4
Django==4.0.6
django-configurations==2.3.2
django-extensions==3.2.0
django-registration==3.3
djangorestframework==3.14.0
filelock==3.7.1
Pillow==9.2.0
platformdirs==2.5.2
pydot==1.4.2
pyparsing==3.0.9
pytz==2022.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
